requirements.txt
flask==3.0.0
flask-cors==4.0.0
gtts==2.5.0
openai>=1.12.0
python-dotenv==1.0.1
httpx>=0.27.0
gunicorn==21.2.0

